changed float to decimal

+++ Assets/Scripts/Entities/ProductInfo.cs
@@ -5,8 +5,8 @@ namespace Entities
     [Serializable]
     public class ProductInfo
     {
-        public string Name;
-        public string Description;
-        public float  Price;
+        public string  Name;
+        public string  Description;
+        public decimal Price;
     }
 }
